refactor(exceptions): use exception handling in kork now that it's there
see https://github.com/spinnaker/kork/pull/821 and https://github.com/spinnaker/spinnaker/issues/5473
