+++ clouddriver-aws/clouddriver-aws.gradle
@@ -23,6 +23,7 @@ dependencies {
   implementation "com.netflix.spinnaker.kork:kork-security"
   implementation "com.netflix.spinnaker.kork:kork-credentials"
   implementation "com.netflix.spinnaker.kork:kork-moniker"
+  implementation "com.netflix.spinnaker.kork:kork-retrofit"
   implementation "com.squareup.okhttp:okhttp"
   implementation "com.squareup.okhttp:okhttp-apache"
   implementation "com.squareup.okhttp:okhttp-urlconnection"
